+++ conformance/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @boardstate/conformance
 
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [[`39083cc`](https://github.com/100yenadmin/boardstate/commit/39083ccdd7b5d5689161b955b37234202467e42b)]:
+  - @boardstate/schema@1.7.0
+  - @boardstate/core@1.7.0
+  - @boardstate/server@1.7.0
+
 ## 0.1.15
 
 ### Patch Changes