+++ examples/standalone/CHANGELOG.md
@@ -1,5 +1,16 @@
 # boardstate-example-standalone
 
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies [[`39083cc`](https://github.com/100yenadmin/boardstate/commit/39083ccdd7b5d5689161b955b37234202467e42b)]:
+  - @boardstate/schema@1.7.0
+  - @boardstate/core@1.7.0
+  - @boardstate/server@1.7.0
+  - @boardstate/host@1.7.0
+  - @boardstate/lit@0.8.0
+
 ## 0.0.16
 
 ### Patch Changes